Extract the PageList component to parallel the newly renamed NavigationJsonPageList.
The new name is marginally better than the old one. Suggestions welcome!

Would recommend writing tests over this code before refactoring so much.
